+++ crates/aspen-jobs/src/manager/config.rs
@@ -17,7 +17,7 @@ pub struct JobManagerConfig {
     /// TTL for deduplication entries.
     pub deduplication_ttl: Duration,
     /// Maximum jobs to schedule per tick.
-    pub max_schedule_per_tick: usize,
+    pub max_schedule_per_tick: u32,
     /// Scheduler tick interval.
     pub scheduler_interval: Duration,
 }
@@ -29,7 +29,7 @@ impl Default for JobManagerConfig {
             default_job_timeout: Duration::from_secs(300),        // 5 minutes
             enable_deduplication: true,
             deduplication_ttl: Duration::from_secs(3600), // 1 hour
-            max_schedule_per_tick: 100,
+            max_schedule_per_tick: 100_u32,
             scheduler_interval: Duration::from_secs(60), // 1 minute
         };
         config.validate();
